cli/diag: Introduce crosshair command
The crosshair command uses the firmware's equivalent MRPC command to
measure the eye diagram.

Like the eye command, there are three output modes: curses, text and
csv. In curses mode the crosshair is plotted on the screen, as it is
measured. Text mode produces a similar text based plot (but only after
the measurement is complete). The CSV mode just dumps the parameters of
the crosshair in a CSV format.

Optionally, the user can specify the -P argument with a CSV file
generated by the eye command to add a background to the crosshair from
the eye capture.